Version from ccuxxx in the secured SHDS folder post unity but pre full annotations
There is no data so it's not expected that this will run end to end given there's nothing to run it on and existing filepaths /dbs it points at won't work. It will however be a close approximation and represent the method. This should make it easier to copy/paste into other notebooks